Build docs.rs documentation with `--all-features`
The docs.rs documentation for `fp-bindgen` is currently built without
all features enabled. This can be confusing, as types that are mentioned
in the README are not present in the API reference. This pull request
should take care of that, by telling docs.rs to pass `--all-features` to
`cargo doc`.

I haven't tested this, but I think it should work. The `docs.rs`
documentation on the topic seems straight-forward enough:
https://docs.rs/about/metadata

Close #120