Fix some warnings (#122)
* Suppress warnings due to deprecated enum variant

* Fix `cargo doc` warnings

The warnings was "this URL is not a hyperlink". `cargo doc` doesn't turn
URLs into links automatically, unless you surround it in `<>`.

There's one of these warnings still left (`generators/mod.rs:75`). Not
sure what to do about that, since the URL documents a value for a
string, and maybe shouldn't become a link in the documentation.

* Update README

The documentation it's generated from was updated in the previous
commit.

* Fix test failures