feat: add RustWasmerWasiRuntime BindingsType (#135)
* feat: add RustWasmerWasiRuntime BindingsType

* refactor: get rid of code duplicates in RustWasmerWasiRuntime

* feat: add RustWasmerWasiRuntime to example-protocol

* fix: pass missing string to enum variant

Recently, the enum variant InvocationError::FunctionNotExported started to expect a string parameter which was added in this commit.

* feat: add example-rust-wasmer-wasi-runtime

* feat: update Rust runtimes to support changes in recent merge

This additional would not be necessary if we had a single example for both wasm32-unknown-unknown and wasm32-wasi. However, for now examples are treated as a kind of demo for newcomers which requires them to be easily understandable. This is achieved by treating example-rust-wasmer-runtime and example-rust-wasmer-wasi-runtime independently.

* refactor: offer wasm32-wasi as a feature for example-rust-wasmer-runtime

* fix: rustfmt on mod files in wasmer generators

* fix: update expected_bindings for wasmer_wasi runtime

* feat: wasmer tests support wasi

* fix: add missing wasm32-wasi target to test workflow

* fix: temporarily add targets for nightly toolchain

* fix: comment in code block in yaml files not allowed